requirements.txt
quart
twscrape
python-dotenv
cachetools
orjson
aiohttp
aiohttp-retry
aiodns
uvloop
hypercorn